% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_project.R
\name{init_project}
\alias{init_project}
\title{Initialize Project}
\usage{
init_project(
  type = c("basic", "modelling", "teradata", "ems", "custom"),
  custom_structure = NA,
  additional_packages = NA
)
}
\arguments{
\item{type}{the kind of project to create - basic, teradata, EMS, or a custom project}

\item{custom_structure}{a list with the custom structure, if used.}

\item{additional_packages}{a string, with additional packages to load along with the template, if any.}
}
\value{
This function is only used for its side effects
}
\description{
Create a initial project structure
}
\details{
Creates a project structure in the GOI style. Creates a basic, modelling, Teradata
or EMS project, or creates a custom one from a custom_structure list.
If a custom structure is needed, you must pass a list with the following elements:
files - a string of filenames to create
folders - a string of folders to create
packages - a string of packages to insert in the first file
additional changes - a list, containing two elements:
files - a list of files to add elements to
text - a string, with the corresponding text to add to each file.
}
\examples{
\dontrun{
init_project("basic")
}
}
