% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{onehot}
\alias{onehot}
\title{Onehot}
\usage{
onehot(condition, treat_na_as = c(NA, 1, 0))
}
\arguments{
\item{condition}{A condition, expressed as var {something} value}

\item{treat_na_as}{a value to turn NA values into - zeros, 1s or NA}
}
\value{
This function returns if x is not na
}
\description{
Returns 1 if condition is met, else zero
}
\details{
It does the opposite of what is.na
}
\examples{
df <- data.frame(x = 1:100)
df \%>\% dplyr::mutate(greater_50_flag = onehot(x > 50))
}
\author{
Matt Simmons mattsimmons@email.com
}
